Move WikiTransformation back to Utils.
