+++ web/php/Actions/PMAction.php
@@ -13,7 +13,7 @@ use Wikidot\Utils\NotificationMaker;
 use Wikidot\Utils\ProcessException;
 use Wikidot\Utils\WDPermissionException;
 use Wikidot\Utils\WDPermissionManager;
-use Wikidot\Wikitext\WikiTransformation;
+use Wikidot\Utils\WikiTransformation;
 
 class PMAction extends SmartyAction
 {
