+++ web/php/Utils/Outdater.php
@@ -17,7 +17,6 @@ use Wikidot\DB\PageInclusionPeer;
 use Wikidot\DB\PageInclusion;
 use Wikidot\DB\CategoryPeer;
 use Wikidot\DB\SitePeer;
-use Wikidot\Wikitext\WikiTransformation;
 
 class Outdater
 {
