+++ web/php/Utils/WDEditUtils.php
@@ -3,7 +3,6 @@
 namespace Wikidot\Utils;
 
 use Ozone\Framework\OzoneLogger;
-use Wikidot\Wikitext\WikiTransformation;
 
 class WDEditUtils
 {
